fix: cross-platform vet errors exposed by FreeBSD support
Three small fixes to make `go vet` clean across all
GOOS/GOARCH combinations Go supports:

- server/fstest/fstest_test.go gets `//go:build linux || freebsd`;
  it imports `passthrough.NewRoot` which is now gated to those two
  OSes
- server/passthrough/stat_linux.go and passthrough_test.go cast
  `st.Nlink` (uint32 on linux/arm64, uint64 elsewhere) to uint64
  for the proto.Attr.NLink assignment / comparison
- proto/proto_test.go casts math.MaxUint32 to uint32 in a t.Errorf
  arg so freebsd/386 (and any other 32-bit target) doesn't overflow
  the implicit `int` conversion

+++ proto/proto_test.go
@@ -471,7 +471,7 @@ func TestConstants(t *testing.T) {
 		t.Errorf("NoTag = %d, want %d", NoTag, math.MaxUint16)
 	}
 	if NoFid != Fid(math.MaxUint32) {
-		t.Errorf("NoFid = %d, want %d", NoFid, math.MaxUint32)
+		t.Errorf("NoFid = %d, want %d", NoFid, uint32(math.MaxUint32))
 	}
 	if QIDSize != 13 {
 		t.Errorf("QIDSize = %d, want 13", QIDSize)
